requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
outcome==1.2.0
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
webdriver-manager==4.0.0
websockets==11.0.3
wsproto==1.2.0
